auth-server: bundle-store: Rename module to `bundle_store`
This is a slightly clearer name that documents *what* is being stored
